Define rota para cadastro de usuário
