+++ grails-app/controllers/com/mini/asaas/customer/CustomerController.groovy
@@ -9,24 +9,7 @@ class CustomerController {
     
     CustomerService customerService
 
-    def index() {
-
-    }
-
-    def save() {
-        try {
-            CustomerAdapter adapter = new CustomerAdapter(params)
-            Customer customer = customerService.save(adapter)
-
-            redirect(action: "show", id: customer.id)
-        } catch (Exception exception) {
-            log.error("CustomerController.save >> Erro ao cadastrar cliente", exception)
-            flash.type = MessageType.ERROR
-            flash.message = 'Erro ao salvar os dados, verifique todos os campos e tente novamente.'
-
-            redirect(action: "index")
-         }
-    }
+    def index() {}
 
     def show() {
         try {
